+++ app/board/main/page.tsx
@@ -1,120 +1,28 @@
-"use client";
-
 import Footer from "@/components/footer";
 import Header from "@/components/header";
-import { Button, Pagination } from "@nextui-org/react";
-import Link from "next/link";
-import { useEffect, useState } from "react";
-import { privateApi } from "@/api/axiosConfig";
-import GetUserInfo from "@/hook/getUserInfo";
-import axios from "axios";
-
-interface Board {
-  boardCid: number;
-  boardName: string;
-}
-
-interface Post {
-  postCid: number;
-}
-
-interface IPostInfo {
-  author: string;
-  createdAt: string;
-  postCid: number;
-  postTitle: string;
-  postView: number;
-}
+import BoardComponent from "@/components/test/boardComponent";
 
 export default function Main() {
-  const [boardData, setBoardData] = useState<IPostInfo[] | null>([]);
-  const [postData, setPostData] = useState(null);
-  const [errorMessage, setErrorMessage] = useState("");
-  const [page, setPage] = useState<number>(1);
-  const [totalPage, setTotalPage] = useState<number>(1);
-
-  //추후에 사용 다른페이지에서
-  const boardCid = 1;
-
-  useEffect(() => {
-    const fetchBoardData = async () => {
-      try {
-        const response = await privateApi.get(
-          `/board/getBoard/${boardCid}/${page}`
-        );
-        // 응답 처리
-        if (response.data.success) {
-          setBoardData(response.data.postList);
-          setErrorMessage("");
-          setTotalPage(response.data.boardInfo.totalPages);
-        }
-      } catch (error) {
-        if (axios.isAxiosError(error)) {
-          setErrorMessage(error.response?.data.message);
-        }
-      }
-    };
-    fetchBoardData(); // 게시판 데이터 불러오기
-  }, [page]);
-
   return (
     <div className="flex h-screen justify-center items-center">
       <div className="max-w-[767px] flex flex-col items-center border-1 border-[#d1d5db] bg-white shadow-lg rounded-lg">
         <Header />
-        <div className="w-96 h-[600px] m-2 p-4 border-1 border-[#d1d5db] bg-white">
-          <main className="pb-2 flex justify-between text-xl tracking-widest items-center border-b-2 border-gray-700 pl-1 pr-1">
-            전체 게시판
-            <Link href="/board/post/create">
-              <Button
-                isIconOnly
-                aria-label="post"
-                className="bg-sub_purple float-right"
-              >
-                <img
-                  src="/icons/post.png"
-                  width="25"
-                  height="25"
-                  style={{ filter: "brightness(0) invert(1)" }} // 이미지 색 white로 변경
-                />
-              </Button>
-            </Link>
-          </main>
-          <div className="h-[450px] overflow-auto scrollbar-none">
-            {errorMessage && <p>{errorMessage}</p>}
-            {boardData &&
-              boardData.map((post) => (
-                <Link
-                  href={`/board/post/read/${post.postCid}`}
-                  key={post.postCid}
-                >
-                  <div className="border-b-1 border-gray-400 pb-2 pt-2 cursor-pointer pl-1">
-                    <div>
-                      <span className="flex text-lg">
-                        {post?.postTitle && post.postTitle.length > 20
-                          ? post.postTitle.slice(0, 20) + "..."
-                          : post?.postTitle}
-                      </span>
-                    </div>
-                    <div className="flex gap-2 text-sm text-gray-500">
-                      <span>{post.author} |</span>
-                      <span>{post.createdAt} |</span>
-                      <span>{post.postView}</span>
-                    </div>
-                  </div>
-                </Link>
-              ))}
-          </div>
-          <Pagination
-            showControls
-            total={totalPage}
-            initialPage={page}
-            className="mt-3 flex justify-center"
-            color="secondary"
-            onChange={(page: number) => setPage(page)}
-          />
-        </div>
+        <BoardComponent boardName={"전체 게시판"} />
         <Footer />
       </div>
     </div>
   );
 }
+
+// 페이지 3개가 동일함
+// -> 페이지의 데이터값만 수정해서 같은 컴포넌트를 사용
+// 게시판 데이터를 불러오기 위해서 필요한것들
+// 1. boardCid -> UserBoard
+// 2. boardName -> UserBoard
+// 3. page -> state값으로 페이지 마다 보관
+// 4. totalPage -> boardInfo
+
+// 상위 컴포넌트 -> 헤더, 푸터, 현재 게시판 정보
+// 메인페이지를 제외하고는 UserBoard정보를 넘겨받을수가 있음 (path값으로 받을수 있음)
+// url에 게시판이름을 넣음 -> url에서 게시판 이름을 꺼내오고 userBoard에 있는 거랑 매치시켜서 cid를 가져옴
+// 커뮤니티 = 커뮤니티 게시판, 2311기수 게시판 = 2311기수 게시판
